Parsing done for mandatory part
