+++ scenes/too_many_elements.rt
@@ -1,7 +1,7 @@
 A 0.2                                   255,255,255
 C -50,0,20        0,0,0,   70
-C -50,0,20        0,0,0,   70
 L -40,0,30                 0.7          255,255,255
 pl 0,0,0          0,1,0                 255,0,225
 sp 0,0,20                  20           255,0,0
 cy 50.0,0.0,20.6  0,0.1,0  14.2  21.42  10,0,255
+L -40,0,30                 0.7          255,255,255
