+++ src/parser/parse_element.c
@@ -6,7 +6,7 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/17 10:58:50 by henbuska          #+#    #+#             */
-/*   Updated: 2025/02/19 14:46:53 by henbuska         ###   ########.fr       */
+/*   Updated: 2025/02/20 13:20:28 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	parse_ambient(char **element, t_rt *rt)
 {
 	double	ambient;
 	char	**colors;
-	
+
 	if (rt->scene->ambient.id != 0)
 		return (error("Too many ambient lights in file", 1));
 	if (validate_argument_count(element, 3))
@@ -28,12 +28,8 @@ int	parse_ambient(char **element, t_rt *rt)
 	if (!colors)
 		return (error("Invalid color for ambient", 1));
 	rt->scene->ambient.ratio = ambient;
-	rt->scene->ambient.color.x = 255.0/custom_atoi(colors[0]);
-	rt->scene->ambient.color.y = 255.0/custom_atoi(colors[1]);
-	rt->scene->ambient.color.z = 255.0/custom_atoi(colors[2]);
+	rt->scene->ambient.color = string_to_color(colors);
 	rt->scene->ambient.id = 1;
-	free_array(colors);
-	printf("Ambient ratio: %f, colors: %f, %f, %f\n", rt->scene->ambient.ratio, rt->scene->ambient.color.x, rt->scene->ambient.color.y, rt->scene->ambient.color.z);
 	return (0);
 }
 
@@ -50,25 +46,16 @@ int	parse_camera(char **element, t_rt *rt)
 	coordinates = validate_coordinates(element[1]);
 	if (!coordinates)
 		return (error("Invalid coordinates for camera view point", 1));
-	rt->scene->camera.view_point.x = ft_atof(coordinates[0]);
-	rt->scene->camera.view_point.y = ft_atof(coordinates[1]);
-	rt->scene->camera.view_point.z = ft_atof(coordinates[2]);
-	free_array(coordinates);
-	printf("Camera view point: %f, %f, %f\n", rt->scene->camera.view_point.x, rt->scene->camera.view_point.y, rt->scene->camera.view_point.z);
+	rt->scene->camera.view_point = string_to_point(coordinates);
 	normal = validate_vector(element[2]);
 	if (!normal)
 		return (error("Invalid orientation vector for camera", 1));
-	rt->scene->camera.normal.x = ft_atof(normal[0]);
-	rt->scene->camera.normal.y = ft_atof(normal[1]);
-	rt->scene->camera.normal.z = ft_atof(normal[2]);
-	free_array(normal);
-	printf("Camera normal vector: %f, %f, %f\n", rt->scene->camera.normal.x, rt->scene->camera.normal.y, rt->scene->camera.normal.z);
-	fov = custom_atoi(element[3]);
+	rt->scene->camera.normal = string_to_vector(normal);
+	fov = rt_atoi(element[3]);
 	if (fov < 0 || fov > 180)
 		return (error("Invalid field of view for camera", 1));
 	rt->scene->camera.field_of_view = fov;
 	rt->scene->camera.id = 1;
-	printf("Camera fow: %f\n", rt->scene->camera.field_of_view);
 	return (0);
 }
 
@@ -77,7 +64,7 @@ int	parse_light(char **element, t_rt *rt)
 	char	**coordinates;
 	char	**colors;
 	double	brightness;
-	
+
 	if (rt->scene->light.id != 0)
 		return (error("Too many lights in file", 1));
 	if (validate_argument_count(element, 4))
@@ -85,32 +72,23 @@ int	parse_light(char **element, t_rt *rt)
 	coordinates = validate_coordinates(element[1]);
 	if (!coordinates)
 		return (error("Invalid coordinates for light point", 1));
-	rt->scene->light.position.x = ft_atof(coordinates[0]);
-	rt->scene->light.position.y = ft_atof(coordinates[1]);
-	rt->scene->light.position.z = ft_atof(coordinates[2]);
-	free_array(coordinates);
-	printf("Light point: %f, %f, %f\n", rt->scene->light.position.x, rt->scene->light.position.y, rt->scene->light.position.z);
+	rt->scene->light.position = string_to_point(coordinates);
 	brightness = validate_ratio(element[2]);
 	if (brightness == -1)
 		return (error("Invalid brightness for light [0.0-1.0]", 1));
 	rt->scene->light.brightness = brightness;
-	printf("Light brightness: %f\n", rt->scene->light.brightness);
 	colors = validate_color(element[3]);
 	if (!colors)
 		return (error("Invalid color for light", 1));
-	rt->scene->light.color.x = 255.0/custom_atoi(colors[0]);
-	rt->scene->light.color.y = 255.0/custom_atoi(colors[1]);
-	rt->scene->light.color.z = 255.0/custom_atoi(colors[2]);
+	rt->scene->light.color = string_to_color(colors);
 	rt->scene->light.id = 1;
-	free_array(colors);
-	printf("Light color: %f, %f, %f\n", rt->scene->light.color.x, rt->scene->light.color.y, rt->scene->light.color.z);
 	return (0);
 }
 
 int	parse_element(char **element, t_rt *rt)
 {
 	if (ft_strcmp(element[0], "A") == 0)
-		return(parse_ambient(element, rt));
+		return (parse_ambient(element, rt));
 	else if (ft_strcmp(element[0], "C") == 0)
 		return (parse_camera(element, rt));
 	else if (ft_strcmp(element[0], "L") == 0)
@@ -120,7 +98,7 @@ int	parse_element(char **element, t_rt *rt)
 	else if (ft_strcmp(element[0], "pl") == 0)
 		return (parse_plane(element, rt));
 	else if (ft_strcmp(element[0], "cy") == 0)
-		return(parse_cylinder(element, rt));
+		return (parse_cylinder(element, rt));
 	else
 	{
 		printf("Invalid element identifier\n");
